separate app setup from server entry point
